Move default `verify_...` fixtures to conftest.py
