+++ testbed/tests/widgets/test_multilinetextinput.py
@@ -22,8 +22,6 @@ from .properties import (  # noqa: F401
     test_placeholder_focus,
     test_readonly,
     test_text_value,
-    verify_focus_handlers,
-    verify_vertical_alignment,
 )
